Extract all things newtypes into their own module
